pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "daily-ai-paper-digest"
version = "0.1.0"
description = "Generate daily research digests for VLA, world models, multimodal LLMs, and agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "httpx>=0.27.0,<1.0.0",
  "jinja2>=3.1.4,<4.0.0",
  "PyYAML>=6.0.2,<7.0.0",
]

[project.scripts]
daily-paper = "daily_paper.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/daily_paper"]

[tool.pytest.ini_options]
testpaths = ["tests"]

